feat: Create Project Collaboration

+++ wondrous-app/components/Common/Filter/index.tsx
@@ -87,7 +87,9 @@ function Filter(props: IFilterProps) {
   };
 
   const handleRemove = (filter, shouldCloseModal = true) => {
-    onRemove(filter);
+    if (onRemove) {
+      onRemove(filter);
+    }
     if (open) {
       setOpen(!shouldCloseModal);
     }
@@ -140,7 +142,7 @@ function Filter(props: IFilterProps) {
           <FilterValues>
             <Icon style={{ backgroundColor: '#0f0f0f', borderRadius: '6px' }} height="26" width="26" />
             {displayNames ? (
-              <InlineText>{`${filterSchema?.label}: ${displayNames}`}</InlineText>
+              <InlineText>{`${filterSchema?.label ? `${filterSchema?.label}: ` : ''} ${displayNames}`}</InlineText>
             ) : (
               filterSchema?.label || 'Filters'
             )}
